Revert "NAS-106381 / 12.0 / get rid of /tmp/failover.json"
